==> Analysis/Data_preparation/1_Sample_populations.ipynb <==
# Preparation of sample metadata

For the scope of this study, we define a "sample" as an individual from whom genomic data has been collected.

Sample metadata, stored in the `Metadata/samples.csv` file, contains information on the African ethnolinguistic classification of each sample, also known as the subpopulation group, and the dataset to which each sample belongs, namely, [GnomAD 1000 Genomes or HGDP](https://gnomad.broadinstitute.org/news/2021-10-gnomad-v3-1-2-minor-release/).

Sample metadata was prepared for analysis by:
1. Selecting and renaming features of interest, namely the sample name and ethnolinguistic classification
2. An additional feature, namely the African continental region from which the ethnolinguistic subpopulation resides, was added to the data

## Import libraries and modules

Change working directory.

import os

os.chdir(
    r"C:\Users\User\Desktop\Megan\MSC2\Results\5._Posthoc_analysis\Pipeline_GnomAD_14032023\Analysis"
)

Import modules and packages.

import sys

sys.path.append(
    r"C:\Users\User\Desktop\Megan\MSC2\Results\5._Posthoc_analysis\Pipeline_GnomAD_14032023"
)
import Utils.constants as constants
import Utils.functions as functions

## Import data

Import data from the Metadata/samples.csv file.

sample_metadata = constants.SAMPLE_INFO_DF
sample_metadata.head(5)

## Feature selection

Select the sample name and subpopulation features for further analysis. Rename features if necessary.

The different African subpopulation groups/ethnolinguistic classifications are: 
* ASW: African American
* ACB: African Caribbean
* GWD: Mandinka
* ESN: Esan
* MSL: Mende
* MbutiPygmy: Mbuti Pygmy
* BiakaPygmy: Biaka Pygmy
* Mandenka
* YRI: Yoruba
* San
* BantuSouthAfrica: Bantu from South Africa
* LWK: Luhya

sample_subpopulations = sample_metadata[["SUB", "sample_name"]].rename(columns={"sample_name": "SAMPLE_NAME"})
sample_subpopulations.head(5)

## Add a feature

Add data on the regional classification of a sample

The different regional groupings are: 

* ASW: African American
* ACB: African Caribbean
* SA: Southern Africa
* WA: Western Africa
* CA: Central Africa
* EA: Eastern Africa

sample_subpopulations[
    "REG"
] = sample_subpopulations["SUB"].map(constants.LD_REGIONAL_CLASSIFICATION)

sample_subpopulations.tail(5)

## Save population count data to CSV file

sample_subpopulations.reset_index(drop=True).to_csv(
    os.path.join(
        constants.HOME_PATH,
        "Data",
        "Processed",
        "Sample_populations.csv",
    )
)
