==> Analysis/Data_preparation/3_ALFA_allele_counts_b.ipynb <==
## ALFA data preparation part B
After retrieving the global variant count data from the NCBI ALFA database, the data must be organized and prepared in a suitable format for further analysis. The following steps were taken to achieve this:

* Unpacking Dictionary Values: Some columns in the data contain information stored in a dictionary format. To make the data more accessible for analysis, these dictionary values were unpacked and expanded into multiple separate columns.
* Translating Coded Population Names: The names of populations have been represented by codes in the retrieved data. To make the data easier to understand, these coded population names were translated into legible and recognizable population names.
* Formatting: To facilitate a meaningful comparison with the in-house allele count data, the retrieved global variant count data was restructured and prepared in a format that aligns with the data format used in the in-house allele count data.

## Import libraries and modules

import os

os.chdir(
    r"C:\Users\User\Desktop\Megan\MSC2\Results\5._Posthoc_analysis\Pipeline_GnomAD_14032023\Analysis"
)

import sys

sys.path.append(
    r"C:\Users\User\Desktop\Megan\MSC2\Results\5._Posthoc_analysis\Pipeline_GnomAD_14032023"
)
import pandas as pd
import Utils.constants as constants
import Utils.functions as functions

## Import retrieved ALFA data

all_alfa_allele_counts = pd.read_csv(
    os.path.join(
        constants.HOME_PATH,
        "Data",
        "Processed",
        "ALFA_allele_counts_a.csv",
    )
).drop(columns="Unnamed: 0")

all_alfa_allele_counts.head(5)

## Unpack dictionary values of the allele count column into multiple separate columns.

# Unpack dictionary of allele count column into multiple separate columns
normalised_count_columns = (
    all_alfa_allele_counts["allele_counts"].map(eval).apply(pd.Series)
)

# Drop the old allele count column from the dataframe
drop_count_column = all_alfa_allele_counts.drop(columns="allele_counts")

# Re-add the new, unpacked allele count column to the dataframe
normalised_alfa_allele_counts = pd.concat(
    [
        drop_count_column.reset_index(drop=True),
        normalised_count_columns.reset_index(drop=True),
    ],
    axis=1,
)

normalised_alfa_allele_counts.head(5)

## Translate population codes into legible population names.

# Fetch ALFA study population metadata
metadata_json = functions.get_metadata()

# Generate a list of all unique study and population codes that need to be queried
study_codes = [
    study_code for study_code in normalised_alfa_allele_counts.study_code.unique()
]
population_codes = [
    population_code
    for population_code in normalised_alfa_allele_counts.population_code.unique()
]

# Fetch study and population name metadata. Code modified from https://github.com/ncbi/dbsnp/blob/master/tutorials/Variation%20Services/Jupyter_Notebook/by_rsid.ipynb.
metadata = {}
for project_json in metadata_json:
    p = {}
    p["json"] = project_json
    p["pops"] = {}
    metadata[project_json["bioproject_id"]] = p
for prj_id, prj in metadata.items():
    functions.add_all_pops(prj["json"]["populations"], prj)

for study_id in study_codes:
    study_name = metadata[study_id]["json"]["short_name"]

    pop_dict = {
        "SAMN10492696": "African Others",
        "SAMN10492698": "African American",
        "SAMN10492697": "East Asian",
        "SAMN10492701": "Other Asian",
    }
    for pop_id in population_codes:
        if pop_id not in [
            "SAMN10492696",
            "SAMN10492698",
            "SAMN10492697",
            "SAMN10492701",
        ]:
            population_name = metadata[study_id]["pops"][pop_id]["name"]
            temp_pop_dict = {pop_id: population_name}
        pop_dict.update(temp_pop_dict)

# Add column in population allele count dataframe for population names
normalised_alfa_allele_counts["Population"] = normalised_alfa_allele_counts[
    "population_code"
].map(pop_dict)

normalised_alfa_allele_counts.head(5)

## Restructure and prepare data in a format that aligns with the data format used in the in-house allele count data.

The data was "melted" to reorganize it in a more understandable format. In the original data, each row might contain counts for multiple alleles corresponding to a single variant. To make the data easier to work with, each row was transformed so that it now represents the count for just one allele. 

melted_alfa_allele_counts = normalised_alfa_allele_counts.melt(
    id_vars=[
        "study_code",
        "variant_id",
        "population_code",
        "reference_allele",
        "Population",
    ],
    value_vars=normalised_alfa_allele_counts.iloc[:, 4:].columns,
    var_name="allele",
    value_name="count",
).dropna(subset="count")

melted_alfa_allele_counts.head(5)

In the format above, the reference allele count is not explicitly stated. To address this issue and provide a more clear representation of the data, the data was reformatted. By making this adjustment, the new data format now explicitly indicates the counts for the reference allele and the alternate allele(s) for each variant. 







# Separate reference allele and allele count information into different dataframes

ref_allele_info = melted_alfa_allele_counts[
    ["study_code", "variant_id", "population_code", "Population", "reference_allele"]
]
allele_count_info = melted_alfa_allele_counts[
    ["study_code", "variant_id", "population_code", "Population", "allele", "count"]
]

# Add in the allele count information corresponding to each reference allele

ref_alfa_counts = (
    ref_allele_info.merge(
        allele_count_info,
        how="left",
        left_on=[
            "study_code",
            "variant_id",
            "population_code",
            "Population",
            "reference_allele",
        ],
        right_on=[
            "study_code",
            "variant_id",
            "population_code",
            "Population",
            "allele",
        ],
    )
    .rename(columns={"count": "alfa_ref_cts"})
    .drop(columns="allele")
    .drop_duplicates(
        subset=[
            "study_code",
            "variant_id",
            "population_code",
            "Population",
            "reference_allele",
        ]
    )
)

# The remaining allele count information must be that of the alternate alleles. Add them in.

alt_alfa_counts = melted_alfa_allele_counts.merge(
    ref_alfa_counts,
    how="left",
    on=[
        "study_code",
        "variant_id",
        "population_code",
        "reference_allele",
        "Population",
    ],
).rename(columns={"allele": "alternate_allele", "count": "alfa_alt_cts"})

alfa_counts = alt_alfa_counts[
    alt_alfa_counts["reference_allele"] != alt_alfa_counts["alternate_allele"]
]

alfa_counts.head(5)

Restructure ALFA data to be in the same format as the in-house variant count data

# Pivot data
alfa_pivot_data = alfa_counts.pivot(
    index=["variant_id", "reference_allele", "alternate_allele"],
    columns="Population",
    values=["alfa_alt_cts", "alfa_ref_cts"],
)

# Separate alternate and reference count data into different dataframes to facilate renaming of count columns appropriately
alfa_data_alt = (
    alfa_pivot_data[["alfa_alt_cts"]].droplevel(level=0, axis=1).reset_index()
)
alfa_data_ref = (
    alfa_pivot_data[["alfa_ref_cts"]].droplevel(level=0, axis=1).reset_index()
)

# Add appropriate prefixes to alt and ref columns
alfa_data_alt = functions.add_prefix_dataframe_col_names(
    alfa_data_alt, alfa_data_alt.iloc[:, 3:], "ALT_CT_ALFA_"
)

alfa_data_ref = functions.add_prefix_dataframe_col_names(
    alfa_data_ref, alfa_data_ref.iloc[:, 3:], "REF_CT_ALFA_"
)

# Merge renamed alternate and reference count data
alfa_grouped_data = alfa_data_alt.merge(
    alfa_data_ref, on=["variant_id", "reference_allele", "alternate_allele"]
)

alfa_grouped_data.head(5)

## Save formatted ALFA results to csv

alfa_grouped_data.reset_index(drop=True).to_csv(
    os.path.join(
        constants.HOME_PATH,
        "Data",
        "Processed",
        "ALFA_allele_counts_b.csv",
    )
)
