==> Analysis/Data_preparation/7_Variant_effects.ipynb <==
# Data Preparation

## Preparation of variant effect data

Change working directory

import os

os.chdir(
    r"C:\Users\User\Desktop\Megan\MSC2\Results\5._Posthoc_analysis\Pipeline_GnomAD_14032023\Analysis"
)

Import modules and packages

import sys

sys.path.append(
    r"C:\Users\User\Desktop\Megan\MSC2\Results\5._Posthoc_analysis\Pipeline_GnomAD_14032023"
)
import pandas as pd
import Utils.constants as constants
import Utils.functions as functions

Import PredictSNP2 variant effect data

pdtsnp2_vep_data = pd.DataFrame()
vep_path = os.path.join(
    constants.HOME_PATH, "Data", "Raw", "PredictSNP2", "PredictSNP2.csv"
)
if os.path.exists(vep_path):
    pdtsnp2_vep_data = pd.read_csv(vep_path, sep=";")

pdtsnp2_vep_data.head(5)

Add gene and positional information

# Import gene and positional data from CSV file
gene_positional_data = pd.read_csv(
    os.path.join(
        constants.HOME_PATH,
        "Data",
        "Processed",
        "Variant_types_and_consequences.csv",
    ),
    usecols=["ID", "REF", "ALT", "POS", "GENE"],
)

# Merge gene and positional data with frequency information
pdtsnp2_vep_data = (
    pdtsnp2_vep_data.merge(
        gene_positional_data,
        how="left",
        on=[
            "POS",
            "REF",
            "ALT",
        ],
    )
    .drop(columns=["ID_x"])
    .rename(columns={"ID_y": "ID"})
)

pdtsnp2_vep_data.head(5)

Save data to csv file

pdtsnp2_vep_data.reset_index(drop=True).to_csv(
    os.path.join(
        constants.HOME_PATH,
        "Data",
        "Processed",
        "Variant_effects.csv",
    )
)
